Update benchmarks to test all the variations.
This permits to have a clear view of all the variations and how they perform,
especially that there is variations and the variations are surprisingly
different across platforms.

Split the code into two support files: benchmark_gpio_support.go and
benchmark_fast_gpio_support.go. This permits two things:

- The exact same code is copy pasted across benchmarks.
- Yet the code is statically using the relevant concrete type, ensuring maximum
  performance. It's like generics, but without the < >.
- The FastXX version are split in a separate file since sysfs's Pin doesn't
  support these functions.

runtime.LockOSThread() is not used as it was not observed to have significant
impact for the benchmark measurements.

These were manually tested on a Raspberry Pi 3, Raspberry Pi Zero Wireless and a
C.H.I.P.