periph: Add Driver.After() optional function.
- Make bcm283x and allwinner use it, so they are serialized after sysfs-gpio is
  done processing.

This forces serialization of the drivers, which will simplify the code *a lot*,
especially with gpioreg: this will permit the GPIO registry to become much more
sane!

Cost in startup perf seems to be non-measurable (too much jitter) and stays in
the <=10ms range on RPi3 and <=60ms range for a C.H.I.P. (slow single core).

Fixes #232