headers drivers: use After instead of Prerequisites
Now that the allwinner and bcm283x Pin can work even if the driver failed by
falling back into sysfs-gpio, the header should load even if the CPU GPIO
drivers failed.

allwinner-gpio-pl is now after sysfs-gpio instead of allwinner-gpio as a
prerequisite.

This will enable cases like running Arch on a RPi3 where gpiomem kernel driver
is not loaded, the process is not running as root, so /dev/mem cannot be memory
mapped, but sysfs-gpio loaded successfully.