sysfs: Fix TxPackets() to behave as expected.
spi: improve documentation accordingly.

A 'fix' was implemented in 1c188d54b4e7 but it was not good for TxPackets(). Fix
the implementation and fix the documentation.

The next step is to fix sysfs SPI implementation to stop refusing large
transactions, and instead transparently split them up. Filed as issue #310.

Fixes #222
(again)