Update import path to what host will be
Also update imports to conn.

Ran (roughly):
  # Update the references.
  git ls-files -- . | xargs -L 1 sed -i 's#x/periph/conn#x/conn#g'
  git ls-files -- . | xargs -L 1 sed -i 's#x/periph/experimental/conn#x/conn#g'
  git ls-files -- . | xargs -L 1 sed -i 's#x/periph/host#x/host#g'
  git ls-files -- . | xargs -L 1 sed -i 's#x/periph/experimental/host#x/host#g'
  git ls-files -- . | xargs -L 1 sed -i 's#x/periph#x/host/hostreg#g'
  git ls-files -- . | xargs -L 1 sed -i 's#\tperiph\.#\thostreg.#g'

  # Initialize the go module:
  go mod init periph.io/x/host

  # Edited it to require go 1.13:
  vi go.mod

  # Edited to update to hostreg
  vi host.go

  # Reorder imports.
  goimports -w .

  # Forced go.mod and go.sum update
  go test ./...

I had more trouble than the above because I initially tried to setup
periph.io/x/conn/v0 but that is not allowed, the problem is that I
realized only after pushing v0.1.0. Anyhow with a forced pull via:
  go get periph.io/x/conn@a1854b91a64c160ebf9ab0d5e52be5cb8c1b7676
I was able to get past the bad commits with the go.mod ending with /v0.